Update Decimal_Precision.h
I have a feeling that all of these re-definitions are not a big deal since we are only using C++. Most of these functions that are redefined work for all types of numbers in C++. If we used C these would become more necessary. I made our special decimal precision variable marb so use that whenever you are working in the marballs namespace.